refactor(executor): split bind.rs into submodules
Split the 1540-line bind.rs file into focused submodules:
- bind/mod.rs (393 lines): Public API and tests
- bind/positional.rs (451 lines): Positional parameter binding (? and $N)
- bind/named.rs (452 lines): Named parameter binding (:name)
- bind/visitor.rs (260 lines): AST traversal for placeholder counting

Benefits:
- Easier to navigate and understand
- Better separation of concerns by binding mode
- Simpler testing per category
- Reduced cognitive load for contributors

Closes #3289
